Movie/requirements.txt
django==3.2
djangorestframework==3.13.1
pandas
zipfile
json
xmltodict
abc